fix(settings): split user directory panels and docs
Summary:
- Extracted the user directory list and recent audit sections into
  dedicated Svelte components so the admin settings files stay within
  the enforced line budgets.
- Added missing OpenAPI descriptions for project_ai_retention request
  fields and refreshed generated API artifacts.

Rationale:
- Keeps the settings surface maintainable while preserving the updated
  admin layout and audit behavior.
- Unblocks backend OpenAPI description coverage checks for Project AI
  retention request payloads.

Tests:
- make openapi-check-ci
- make frontend-api-audit-check
- make web-install
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec prettier --check --log-level warn .
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec eslint . --quiet
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH node
  web/scripts/check-file-budgets.mjs
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH node
  web/scripts/check-dependency-boundaries.mjs
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH node
  web/scripts/check-mobile-route-gates.mjs
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec svelte-check --tsconfig ./tsconfig.json --output machine
  --threshold error
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec vitest run --reporter=dot
  src/lib/features/ticket-detail/run-transcript-real-samples.test.ts
  src/lib/features/ticket-detail/run-transcript-live-updates.test.ts
  src/lib/features/ticket-detail/run-transcript.test.ts
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec vite build --logLevel warn
- PATH=$HOME/.nvm/versions/node/v22.22.0/bin:$PATH corepack pnpm --dir
  web exec playwright test --config <temp-copy> --reporter=dot --quiet
  with PLAYWRIGHT_WEB_PORT=4273
- OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest
  OPENASE_GO_TEST_PROGRESS_MODE=json
  OPENASE_BACKEND_TEST_GROUP_SIZE=8 make check
- OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest
  OPENASE_GO_TEST_PROGRESS_MODE=json make build
- OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest
  OPENASE_GO_TEST_PROGRESS_MODE=json make
  LINT_BASE_REV=$(git merge-base origin/main HEAD) lint
- OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest
  OPENASE_GO_TEST_PROGRESS_MODE=json make lint-depguard
- OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest
  OPENASE_GO_TEST_PROGRESS_MODE=json make lint-architecture
